+++ src/lib/components/ExplorerLink.tsx
@@ -1,10 +1,10 @@
 import type { FlexProps, TextProps } from "@chakra-ui/react";
-import type { AddressReturnType } from "lib/app-provider";
 import type { Option } from "lib/types";
 import type { ReactNode } from "react";
 
 import { Flex, Text } from "@chakra-ui/react";
 import { trackMintScan } from "lib/amplitude";
+import { type AddressReturnType, useChainConfigs } from "lib/app-provider";
 import { useCelatoneApp } from "lib/app-provider/contexts";
 import { useWasmConfig } from "lib/app-provider/hooks/useConfig";
 import { useCurrentChain } from "lib/app-provider/hooks/useCurrentChain";
@@ -28,8 +28,9 @@ export type LinkType =
   | "tx_hash"
   | AddressReturnType;
 
-interface ExplorerLinkProps extends FlexProps {
+export interface ExplorerLinkProps extends FlexProps {
   ampCopierSection?: string;
+  chainId?: string;
   copyValue?: string;
   externalLink?: string;
   fixedHeight?: boolean;
@@ -115,13 +116,23 @@ const getValueText = (
   return isTruncate ? truncate(value) : value;
 };
 
-const getCopyLabel = (type: LinkType) =>
-  type
+const getCopyLabel = (type: LinkType, value: string) => {
+  if (type === "user_address") return "address";
+
+  if (type === "nft_collection") {
+    if (value.includes("/nft/")) {
+      return "nft";
+    }
+  }
+
+  return type
     .split("_")
-    .map((str: string) => str.charAt(0).toUpperCase() + str.slice(1))
+    .map((str: string) => str.charAt(0) + str.slice(1))
     .join(" ");
+};
 
 const LinkRender = ({
+  chainId,
   fallbackValue,
   hrefLink,
   isEllipsis,
@@ -131,6 +142,7 @@ const LinkRender = ({
   textVariant,
   type,
 }: {
+  chainId?: string;
   fallbackValue: string;
   hrefLink: string;
   isEllipsis: boolean;
@@ -157,6 +169,7 @@ const LinkRender = ({
   return isInternal && !openNewTab ? (
     <AppLink
       style={{ overflow: "hidden" }}
+      chainId={chainId}
       href={hrefLink}
       passHref
       onClick={(e) => e.stopPropagation()}
@@ -167,7 +180,7 @@ const LinkRender = ({
     <a
       style={{ overflow: "hidden" }}
       data-peer
-      href={isInternal ? `/${currentChainId}${hrefLink}` : hrefLink}
+      href={isInternal ? `/${chainId ?? currentChainId}${hrefLink}` : hrefLink}
       rel="noopener noreferrer"
       target="_blank"
       onClick={(e) => {
@@ -182,6 +195,7 @@ const LinkRender = ({
 
 export const ExplorerLink = ({
   ampCopierSection,
+  chainId,
   copyValue,
   externalLink,
   fixedHeight = true,
@@ -197,13 +211,14 @@ export const ExplorerLink = ({
   ...componentProps
 }: ExplorerLinkProps) => {
   const isMobile = useMobile();
+  const { chainConfigs } = useChainConfigs();
   const { address } = useCurrentChain();
   const { enabled: wasmEnabled } = useWasmConfig({ shouldRedirect: false });
 
   const [internalLink, textValue] = [
     getNavigationUrl({
       type,
-      value: copyValue || value,
+      value,
       wasmEnabled,
     }),
     textLabel ??
@@ -211,14 +226,28 @@ export const ExplorerLink = ({
   ];
 
   const link = externalLink ?? internalLink;
-  const readOnly = isReadOnly || !link;
+  const isNotInitiaChainId = chainId && !chainConfigs[chainId];
+  const readOnly = isReadOnly || !link || isNotInitiaChainId;
+
   // TODO: handle auto width
   return readOnly ? (
-    <Flex alignItems="center" gap={1} {...componentProps}>
+    <Flex
+      className="copier-wrapper"
+      alignItems="center"
+      gap={1}
+      {...componentProps}
+    >
       <Text color="text.disabled" variant="body2">
         {textValue}
       </Text>
       {rightIcon}
+      <Copier
+        display={showCopyOnHover && !isMobile ? "none" : "inline"}
+        hoverLabel={`Copy ${getCopyLabel(type, value)}`}
+        ml={1}
+        type={type}
+        value={copyValue ?? value}
+      />
     </Flex>
   ) : (
     <Flex
@@ -235,7 +264,8 @@ export const ExplorerLink = ({
       {...componentProps}
     >
       <LinkRender
-        fallbackValue={copyValue || value}
+        chainId={chainId}
+        fallbackValue={copyValue ?? ""}
         hrefLink={link}
         isEllipsis={textFormat === "ellipsis"}
         isInternal={isUndefined(externalLink)}
@@ -247,11 +277,11 @@ export const ExplorerLink = ({
       {rightIcon}
       <Copier
         amptrackSection={ampCopierSection}
-        copyLabel={copyValue ? `${getCopyLabel(type)} Copied!` : undefined}
         display={showCopyOnHover && !isMobile ? "none" : "inline"}
+        hoverLabel={`Copy ${getCopyLabel(type, value)}`}
         ml={1}
         type={type}
-        value={copyValue || value}
+        value={copyValue ?? value}
       />
     </Flex>
   );