+++ src/lib/pages/tx-details/components/TxInfoMobile.tsx
@@ -5,13 +5,13 @@ import type { Option, Ratio } from "lib/types";
 import { chakra, Flex, Grid, GridItem, Text } from "@chakra-ui/react";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { LabelText } from "lib/components/LabelText";
+import { TokenImageWithAmount } from "lib/components/token";
 import { useAssetInfos } from "lib/services/assetService";
 import { useMovePoolInfos } from "lib/services/move/poolService";
 import {
   computeCosmosFee,
   formatInteger,
   formatPrettyPercent,
-  formatTokenWithValue,
 } from "lib/utils";
 
 interface TxInfoMobileProps extends FlexProps {
@@ -77,14 +77,14 @@ export const TxInfoMobile = ({
       <Flex>
         <LabelText flex={1} label="Transaction fee">
           {feeToken ? (
-            formatTokenWithValue(feeToken)
+            <TokenImageWithAmount boxSize={4} token={feeToken} />
           ) : (
             <Text color="text.dark" variant="body2">
               No fee
             </Text>
           )}
         </LabelText>
-        <LabelText flex={1} label="Gas used/wanted">
+        <LabelText flex={1} label="Gas used/requested">
           {`${formatInteger(txData.gasUsed)}/${formatInteger(
             txData.gasWanted
           )}`}