+++ CHANGELOG.md
@@ -44,11 +44,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+### Bug fixes
+
+## v1.14.0
+
+### Features
+
+- [#1360](https://github.com/alleslabs/celatone-frontend/pull/1360) Support Initia API for getting asset info with price
+- [#1356](https://github.com/alleslabs/celatone-frontend/pull/1356) Support get ipfs image from glyph library
+
+### Improvements
+
 - [#1353](https://github.com/alleslabs/celatone-frontend/pull/1353) Update initiajs snippet to match the latest version
 
 ### Bug fixes
 
+- [#1369](https://github.com/alleslabs/celatone-frontend/pull/1369) Fix token card name spacing
+- [#1367](https://github.com/alleslabs/celatone-frontend/pull/1367) Fix network card search params replacement
 - [#1359](https://github.com/alleslabs/celatone-frontend/pull/1359) Remove unwanted 0 prefix for inactive validator
+- [#1357](https://github.com/alleslabs/celatone-frontend/pull/1357) Fix validate uint string rule
 
 ## v1.13.0